Move MySql specific files into own package
So mysql driver isn't required by other sql providers
and other providers can test their eventsource.Store implementation
